[docs/graphql] add information about using curl with GraphQL
## Description 

Adds a brief paragraph in the GraphQL page in our docs website showing
how to query the GraphQL instance through `curl`.

## Test Plan 

Made sure the `curl` command works. 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Improved the GraphQL docs by adding a short paragraph on how to use
`curl` to query the GraphQL service.

---------