[mysticeti] wire in mysticeti client
## Description 

This PR:
* introducing the `LazyMysticetiClient` to be used on the
`consensus_adapter`
* Sharing the `LazyMysticetiClient` between the consensus adapter and
the mysticeti manager to directly update the underneath client
`MysticetiClient` on every epoch change. That was preferred to the
global variable approach as it's more straightforward when needing to
spin up programmatically multiple haneul nodes.
* `SimpleBlockHandler` moved to `Mysticeti` as it was finally proved to
be more convenient that way
* is updating the mysticeti ref

Next:

- [ ] [wire mysticeti consensus handler & boot consensus protocol based
on property](https://github.com/GeunhwaJeong/haneul/pull/14699)

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes