Integrate http based downloader with archival, snapshot and haneul tool
## Description 

Allow using no sign requests for haneul tool, archival and snapshot use
cases in haneul. User just passes in so sign request as a parameter in
their object store config and right downloader tool is used.
Also cleaned up the abstractions and created new ones so in the future
we can support unauthenticated list requests if need be.