Add a simple MANIFEST file to db snapshots
## Description 

For http-based snapshot downloads we need to have an easily accessible
list of the files available in a snapshot. This PR creates a MANIFEST
file which is written to the epoch dir root, and contains the list of
all files available.

## Test Plan 

Just a unit test

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes