Isolate docs site pnpm lockfile
## Description 

I noticed the docs site pnpm lockfile isn't actually fully isolated from
the rest of the repo, but it seems like that's the intent, so I updated
it so that it treats it like its own pnpm root. I also think eventually
we may want to not have these be separate and pull the docs site into
the normal pnpm workspace but leaving that for later.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes