Fix analytics processor and parquet writer integration for empty files
## Description 

Parquet writer does not write a file when called with `flush()` if there
are no records to write:
```
    fn flush(&mut self, end_checkpoint_seq_num: u64) -> Result<bool> {
        if self.data.is_empty() {
            return Ok(false);
        }
```
This is not handled in analytics processor today (with its integration
with csv writer) because it expects the file to be written but later an
empty file is just discarded when uploading to remote store. This PR
fixes analytics processor to handle this scenario properly i.e handle
the scenario when file is not written to disk (as parquet writer does
it).
We basically change the function `flush()` to return a boolean true if
it did write a file or false otherwise so analytics processor can handle
it properly.

## Test Plan 

Ran locally for move-call pipeline (which generates empty files) and
after this fix things work again.