[dapp-kit] update docs to include a snippet about importing the CSS stylesheet
## Description 

A builder in Discord was confused about why the provided UI components
weren't displaying correctly in their app, and it was because they
didn't realize they needed to import the CSS file for dApp Kit into
their application. I noticed we don't have anything in our documentation
about this, so I'm adding a section in the docs with this PR 😄

<img width="1379" alt="image"
src="https://github.com/GeunhwaJeong/haneul/assets/7453188/cf346843-1fbf-4893-9cbd-c0c9c01ad140">

https://haneullabs-labs.slack.com/archives/C0393DWJD7X/p1700380725599999

## Test Plan 
- 👁️ 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes