[cli] remove the 20 bytes address description
## Description 

Fixes the description of the `haneul client merge-coin` command, by
removing the reference to the 20 bytes address. This should close
.

## Test Plan 

Just a CLI help description update, no tests are needed.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Fixed the description of the `haneul client merge-coin` command, by
removing the reference to the old-style 20 bytes address.