Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/dapp-kit@0.10.0

### Minor Changes

- 833821720: Add new useHaneulClientQueries hook to support variable client
queries

## haneullabs/create-dapp@0.1.5

### Patch Changes

-   Updated dependencies [833821720]
    -   haneullabs/dapp-kit@0.10.0

## haneullabs/enoki@0.0.4

### Patch Changes

- 5eb2deeb4: Make enoki flow session observable. Expose state parameter
on useAuthCallback.

## haneullabs/sdk-docs@1.0.9

### Patch Changes

-   Updated dependencies [833821720]
    -   haneullabs/dapp-kit@0.10.0