Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/create-dapp@0.1.6

### Patch Changes

-   Updated dependencies [a67b6ba8d]
    -   haneullabs/dapp-kit@0.10.1

## haneullabs/dapp-kit@0.10.1

### Patch Changes

- a67b6ba8d: Improve the reliability of the `useAutoConnectWallet` hook.

## haneullabs/enoki@0.0.5

### Patch Changes

-   Updated dependencies [7b78284d2]
    -   haneullabs/zklogin@0.3.8

## haneullabs/zklogin@0.3.8

### Patch Changes

-   7b78284d2: Add simple length checks