Enable effects v2 on mainnet
## Description 

Effects v2 has been running on testnet and working well so far.
This PR enables effects v2 on mainnet.

## Test Plan 

devnet + testnet soak

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [x ] breaking change for FNs (FN binary must upgrade)
- [x ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
This PR changes how we store transaction effects on node databases. The
new structure is cleaner and more efficient, also allows for better
implementation in the execution and other layers. This should not
affects SDKs because we translate it to SDK effects types at json rpc
layer. Detailed design is described in
https://github.com/GeunhwaJeong/haneul/issues/12813.