[cli] Fix bug where the address is not decoded from its PublicKey
## Description 

The HaneulAddress should be decoded from `PublicKey` and not `HaneulKeyPair`,
as the alias file encodes the public key as base64.

## Test Plan 

Added new test

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Fixed a bug where reading the CLI alias file resulted in creating the wrong
address.