Do not always seek to first when creating SafeIterator
## Description 

Similar change to  , to SafeIterator

## Test Plan 

Existing unit tests have good coverage for SafeIterator (not for the old
iterator though). I'll send out a test specific PR for the old iterator
next.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes