[move-ide] Fixed doc comment treatment in symbolicator
## Description 

This PR fixes a problem with doc comments being incorrectly extracted
for display in "type-on-hover" IDE feature. We were trying to add doc
comments for identifiers that should never have them. It's more in sync
with how rust-analyzer is extracting and displaying them.

## Test Plan 

Tests have been modified to reflect this and new tests have been added.