Some clean up around ReconfigState guard
## Description 

This is a clean up after going through whether it is possible that the
lock on ReconfigState can starve/deadlock. Based on what I read, as long
as the DB access to update per epoch tables has bounded return time, the
locks shouldn't be starve.

The only other lock that may interact with ReconfigState lock is the
lock on pending_consensus_certificates. We always acquire ReconfigState
lock before acquire pending_consensus_certificates lock.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes