[bridge 10/n] storage
## Description 

Add storage stuff. The meat is in `storage.rs` and `orchestrator.rs`.
Note we haven't wired it up in `fn run_eth_watcher`, because some
modification are needed in the receiving channel, which will come in the
next PR.

High level it works in this way:
```
#[derive(DBMapUtils)]
pub struct BridgeOrchestratorTables {
    /// pending BridgeActions that orchestrator received but not yet executed
    pub(crate) pending_actions: DBMap<BridgeActionDigest, BridgeAction>,
    /// module identifier to starting transaction digest
    pub(crate) haneul_syncer_cursors: DBMap<Identifier, TransactionDigest>,
    /// contract address to starting block
    pub(crate) eth_syncer_cursors: DBMap<ethers::types::Address, u64>,
}
```

1. Orchestrator receives events from syncers. If an event is an action
(e.g. token transfer), then it writes this action to `pending_actions`
(WAL log), then passes it to the executor (coming in a later PR).
2. without waiting for the action to be finished, it updates the cursor
in `haneul_syncer_cursors` or `eth_syncer_cursors`
3. the executor will remove the action from the logs after it's
confirmed to be done.

### Considerations
**Crash Recovery**
* If the program crashes after it is written to WAL and before the
action getting finalized, pending actions still remain in the WAL and
will be picked up during start (to be implemented), even though the
cursor may have progressed further.
* If the program crashes before the action is written into WAL, then
once it starts, the syncer will get this action and send it over again.

**DB writes**
We write `pending_actions` and `xyz_syncer_cursors` in two db txes and
the order is important. We want to persist changes to `pending_actions`
before passing the action to Executor because if for some reason
executor completes the action and remove the entry earlier than the
addition, it will cause the action to remain in the pending log longer
than necessarily (until next restart). It's not end of world but better
to avoid.

**Cursor**
1. Haneul's cursor is TransactionDigest. Today we paginate event query
results with EventID (TransactionDigest + EventIdx). In `haneul_syncer.rs`
we implement the query such that events of a transaction never appear
partially in a page. This enables downstream to track per transaction
easily.
2. Eth's cursor is BlockNumber. Similarly `eth_syncer.rs` guarantees to
return complete list of matched events by block level.

## Test Plan 

unit tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes