[Haneul] run cheap transaction validity checks early
## Description 

Run cheap transaction validity checks when starting to process
transactions and certificates.

The checks include the input size count check, that needs to run before
system overload and input object checks. This check is moved into
`validity_check()` for `ProgrammableTransaction`.

The assumption is that calling `check_version_supported()` and
`validity_check()` contain no cryptography or rocksdb ops, so should be
cheap enough to run multiple times during the processing of a
transaction.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes