[move-ide] Fixed and extended on-hover for struct fields and locals
## Description 

This PR does three related things:

1. Fixes and extends on-hover functionality for struct fields

Not only on-hover information for fields was very simple (just the
type), it was also incorrect in field access chains as the type
displayed on-hover was incorrectly obtained from the access chain
expression (prefix part). For example, old on-hover looked as follows
(not that the type `SomeStruct` rather than `OuterStruct`):


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/cd7ffff6-fd7b-40d9-8857-46f6f52bd4da)

The new on-hover looks as follows, and closely resembles what
rust-analyzer is displaying (field name and type but also type of the
enclosing struct):


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/654a1b67-598f-438c-b790-48a186067b1a)

2. Fixes and extends on-hover functionality for local variables
(including params)

With locals, only type was being displayed on hover, and this type was
not always correct as it was obtained from the use rather than from the
definition itself. For example, old on-hover looked as follows (not that
type of `tmp` should simply be `u64` rather than `&mut u64`):


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/26341124-dffc-4862-ab73-ce3253945e30)

The on-hover looks as follows and closely resembles what rust-analyzer
is displaying (variable name and type, preceded by `let` but only if the
definition is of `let tmp = ...` shape - no `let` for parameter
declarations or for variables declared via unpack):


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/e3994c93-bbd5-4cc0-888f-fe680c6c7d8e)

3. Renames some identifiers to reflect recent changes (in particular,
`ident_type` to `on_hover` as it no longer simply represents a type but
rather information that is being displayed on hover)

## Test Plan 

Tests have been updated to reflect new functionality