[move-ide] Enhancement to on hover for constants
## Description 

This PR extends on-hover functionality for constants. Previously, we
were only displaying constant typo on hover, for example:


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/58813caa-b02a-4b03-8607-36a9429d5400)

With this PR, we display constant name, type, and value, which closely
resembles what rust-analyzer does:


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/b25217fa-ab63-4c5e-be26-911902617f44)

We need custom to-string conversions as those provided by the compiler
for the AST do not look quite right.

## Test Plan 

Existing tests have been adjusted and tests have been added.