[gql][feature] haneulnsRegistration connection
## Description 

Looks like the NameServiceName node is not fully defined in the draft
schema. Do we want this to follow the standard we've set for other
"special" move objects like Coin, CoinMetadata, and StakedHaneul, or are we
fine with having it represented as a string?

If the former, the query syntax would be something like
```
{
  address(
    address: "0x"
  ) {
    nameServiceNames(first: 1) {
      nodes {
        asMoveObject {
          contents {
            json
            type {
              repr
            }
            data
          }
        }
        domain
      }
    }
  }
}
```


## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes