move-package: add name resolution hook
## Description

Adds a `custom_resolve_pkg_name` hook in order to allow for the
dependency graph to discriminate between packages based on a dynamically
resolved symbol (returned by the hook) rather than just package name.
When the hook isn't defined it returns the package manifest name when so
this doesn't affect Haneul package resolution currently. This will make it
possible to consolidate source and on-chain packages (which don't have a
name) within a single graph.

This is part of the work to enable compiling against on-chain
dependencies https://github.com/GeunhwaJeong/haneul/pull/14178.

## Test Plan 

Added unit tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes