[grahpql] Move SystemState into Epoch
## Description 

Moves the system state summary unique fields into the Epoch struct. It
removes `validatorSet` from `SystemStateSummary` as this is part of the
`Epoch` anyway, and it will be retrieved correctly according to the
given epoch.

Supersedes .

## Test Plan 

Used existing tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes