[CI] Run split cluster check on pull requests
## Description 

Introduces a workflow to be run on new pull requests. Workflow calls
into a script which spins up a local cluster of 4 nodes on the host,
with 2 nodes running mainnet release branch and 2 nodes running a
release candidate (in this case, the PR branch head). The new cluster is
regenesis'ed and overridden to identify as mainnet, therefore it can
provide merge-blocking for changes that could introduce backwards
incompatibility or other issues.

If run on merges to main, will also bisect if this fails

## Test Plan 

https://github.com/GeunhwaJeong/haneul/actions/runs/7428260893

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes