Add tests for handle_certificate
This PR is for item 67/68 in the fixit doc. There are multiple
protections in place for incorrect epoch across safe client, stake
aggregator & signature verifier, added some tests to ensure they are not
moved. Added some tests as well for bad user input into handle
certificate. This is already tested for handle transaction.